refactor: convert purge.js to ts
